+++ src/commands.ts
@@ -23,6 +23,7 @@ import { OpenResourceHead } from "./commands/openResourceHead";
 import { Patch } from "./commands/patch";
 import { PatchAll } from "./commands/patchAll";
 import { PatchChangeList } from "./commands/patchChangeList";
+import { PickCommitMessage } from "./commands/pickCommitMessage";
 import { PromptAuth } from "./commands/promptAuth";
 import { PromptRemove } from "./commands/promptRemove";
 import { PullIncommingChange } from "./commands/pullIncomingChange";
@@ -36,6 +37,7 @@ import { Resolved } from "./commands/resolved";
 import { Revert } from "./commands/revert";
 import { RevertAll } from "./commands/revertAll";
 import { RevertChange } from "./commands/revertChange";
+import { RevertExplorer } from "./commands/revertExplorer";
 import { RevertSelectedRanges } from "./commands/revertSelectedRanges";
 import { SwitchBranch } from "./commands/switchBranch";
 import { Update } from "./commands/update";
@@ -84,4 +86,6 @@ export function registerCommands(model: Model, disposables: Disposable[]) {
   disposables.push(new OpenChangeHead());
   disposables.push(new OpenHeadFile());
   disposables.push(new RevertAll());
+  disposables.push(new PickCommitMessage());
+  disposables.push(new RevertExplorer());
 }