+++ src/model.ts
@@ -14,11 +14,10 @@ import {
   ConstructorPolicy,
   IModelChangeEvent,
   IOpenRepository,
-  RepositoryState,
-  Status
+  RepositoryState
 } from "./common/types";
 import { debounce } from "./decorators";
-import { exists, readdir, stat } from "./fs";
+import { readdir, stat } from "./fs";
 import { configuration } from "./helpers/configuration";
 import { RemoteRepository } from "./remoteRepository";
 import { Repository } from "./repository";
@@ -30,6 +29,7 @@ import {
   filterEvent,
   IDisposable,
   isDescendant,
+  isSvnFolder,
   normalizePath
 } from "./util";
 import { matchAll } from "./util/globMatch";
@@ -85,6 +85,13 @@ export class Model implements IDisposable {
       }
       return this;
     })() as unknown) as Model;
+  }
+
+  public openRepositoriesSorted(): IOpenRepository[] {
+    // Sort by path length (First external and ignored over root)
+    return this.openRepositories.sort(
+      (a, b) => b.repository.workspaceRoot.length - a.repository.workspaceRoot.length
+    );
   }
 
   private onDidChangeConfiguration(): void {
@@ -157,7 +164,7 @@ export class Model implements IDisposable {
   @debounce(500)
   private eventuallyScanPossibleSvnRepositories(): void {
     for (const path of this.possibleSvnRepositoryPaths) {
-      this.tryOpenRepository(path);
+      this.tryOpenRepository(path, 1);
     }
 
     this.possibleSvnRepositoryPaths.clear();
@@ -172,6 +179,19 @@ export class Model implements IDisposable {
     }
 
     repository.statusExternal
+      .map(r => path.join(repository.workspaceRoot, r.path))
+      .forEach(p => this.eventuallyScanPossibleSvnRepository(p));
+  }
+
+  private scanIgnored(repository: Repository): void {
+    const shouldScan =
+      configuration.get<boolean>("detectIgnored") === true;
+
+    if (!shouldScan) {
+      return;
+    }
+
+    repository.statusIgnored
       .map(r => path.join(repository.workspaceRoot, r.path))
       .forEach(p => this.eventuallyScanPossibleSvnRepository(p));
   }
@@ -220,34 +240,9 @@ export class Model implements IDisposable {
       return;
     }
 
-    let isSvnFolder = false;
+    const checkParent = level === 0;
 
-    try {
-      isSvnFolder = await exists(path + "/.svn");
-    } catch (error) {
-      // error
-    }
-
-    // If open only a subpath.
-    if (!isSvnFolder && level === 0) {
-      const pathParts = path.split(/[\\/]/);
-      while (pathParts.length > 0) {
-        pathParts.pop();
-        const topPath = pathParts.join("/") + "/.svn";
-
-        try {
-          isSvnFolder = await exists(topPath);
-        } catch (error) {
-          // error
-        }
-
-        if (isSvnFolder) {
-          break;
-        }
-      }
-    }
-
-    if (isSvnFolder) {
+    if (isSvnFolder(path, checkParent)) {
       // Config based on folder path
       const resourceConfig = workspace.getConfiguration("svn", Uri.file(path));
 
@@ -315,11 +310,13 @@ export class Model implements IDisposable {
     return RemoteRepository.open(this.svn, uri);
   }
 
-  public getRepository(hint: any) {
+  public getRepository(hint: any): Repository | null {
     const liveRepository = this.getOpenRepository(hint);
     if (liveRepository && liveRepository.repository) {
       return liveRepository.repository;
     }
+
+    return null;
   }
 
   public getOpenRepository(hint: any): IOpenRepository | undefined {
@@ -342,7 +339,7 @@ export class Model implements IDisposable {
     }
 
     if (hint instanceof Uri) {
-      return this.openRepositories.find(liveRepository => {
+      return this.openRepositoriesSorted().find(liveRepository => {
         if (
           !isDescendant(liveRepository.repository.workspaceRoot, hint.fsPath)
         ) {
@@ -355,6 +352,15 @@ export class Model implements IDisposable {
             external.path
           );
           if (isDescendant(externalPath, hint.fsPath)) {
+            return false;
+          }
+        }
+        for (const ignored of liveRepository.repository.statusIgnored) {
+          const ignoredPath = path.join(
+            liveRepository.repository.workspaceRoot,
+            ignored.path
+          );
+          if (isDescendant(ignoredPath, hint.fsPath)) {
             return false;
           }
         }
@@ -378,19 +384,28 @@ export class Model implements IDisposable {
     return undefined;
   }
 
-  public async getRepositoryFromUri(uri: Uri) {
-    for (const liveRepository of this.openRepositories) {
+  public async getRepositoryFromUri(uri: Uri): Promise<Repository | null> {
+
+    for (const liveRepository of this.openRepositoriesSorted()) {
       const repository = liveRepository.repository;
+
+      // Ignore path is not child (fix for multiple externals)
+      if (!isDescendant(repository.workspaceRoot, uri.fsPath)) {
+        continue;
+      }
 
       try {
         const path = normalizePath(uri.fsPath);
-        const info = await repository.info(path);
+
+        await repository.info(path);
 
         return repository;
       } catch (error) {
-        console.error();
+        // Ignore
       }
     }
+
+    return null;
   }
 
   private open(repository: Repository): void {
@@ -410,8 +425,10 @@ export class Model implements IDisposable {
 
     const statusListener = repository.onDidChangeStatus(() => {
       this.scanExternals(repository);
+      this.scanIgnored(repository);
     });
     this.scanExternals(repository);
+    this.scanIgnored(repository);
 
     const dispose = () => {
       disappearListener.dispose();