+++ src/test/repository.test.ts
@@ -76,7 +76,7 @@ suite("Repository Tests", () => {
   });
 
   test("Try get current branch name", async () => {
-    const repository: Repository | undefined = model.getRepository(
+    const repository: Repository | null = model.getRepository(
       checkoutDir.fsPath
     );
     if (!repository) {
@@ -89,7 +89,7 @@ suite("Repository Tests", () => {
 
   test("Try commit file", async function() {
     this.timeout(60000);
-    const repository: Repository | undefined = model.getRepository(
+    const repository: Repository | null = model.getRepository(
       checkoutDir.fsPath
     );
     if (!repository) {
@@ -110,7 +110,7 @@ suite("Repository Tests", () => {
     assert.equal(repository.changes.resourceStates.length, 1);
 
     const message = await repository.commitFiles("First Commit", [file]);
-    assert.ok(/Committed revision (.*)\./i.test(message));
+    assert.ok(/1 file commited: revision (.*)\./i.test(message));
 
     assert.equal(repository.changes.resourceStates.length, 0);
 
@@ -138,7 +138,7 @@ suite("Repository Tests", () => {
 
     await model.tryOpenRepository(newCheckoutDir.fsPath);
 
-    const newRepository: Repository | undefined = model.getRepository(
+    const newRepository: Repository | null = model.getRepository(
       newCheckoutDir.fsPath
     );
     if (!newRepository) {