+++ src/util.ts
@@ -126,10 +126,10 @@ export function hasSupportToDecorationProvider() {
 
 try {
   const fake = {
-    onDidChangeDecorations: (value: any): any => toDisposable(() => {}),
-    provideDecoration: (uri: any, token: any): any => {}
+    onDidChangeDecorations: (_value: any): any => toDisposable(() => {}),
+    provideDecoration: (_uri: any, _token: any): any => {}
   };
-  const disposable = window.registerDecorationProvider(fake);
+  window.registerDecorationProvider(fake);
   hasDecorationProvider = true;
   // disposable.dispose(); // Not dispose to prevent: Cannot read property 'provideDecoration' of undefined
 } catch (error) {}
@@ -198,4 +198,23 @@ export function fixPegRevision(file: string) {
   }
 
   return file;
+}
+
+export async function isSvnFolder(dir: string, checkParent: boolean = true): Promise<boolean> {
+
+  const result = await exists(`${dir}/.svn`);
+
+  if (result || !checkParent) {
+    return result;
+  }
+
+  const parent = path.dirname(dir);
+
+  // For windows: the `path.dirname("c:")` return `c:`
+  // For empty or doted dir, return "."
+  if (parent === dir || parent === ".") {
+    return false;
+  }
+
+  return await isSvnFolder(parent, true);
 }